+++ SecondYearHomeworks.sln
@@ -3,6 +3,10 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{F4E8263B-9AEF-417B-AE24-18B083718140}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MatrixMul", "src\MatrixMul\MatrixMul.csproj", "{91C6E107-74A9-456B-B983-ACACA8ABB394}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -10,5 +14,14 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{91C6E107-74A9-456B-B983-ACACA8ABB394}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{91C6E107-74A9-456B-B983-ACACA8ABB394}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91C6E107-74A9-456B-B983-ACACA8ABB394}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{91C6E107-74A9-456B-B983-ACACA8ABB394}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{91C6E107-74A9-456B-B983-ACACA8ABB394} = {F4E8263B-9AEF-417B-AE24-18B083718140}
 	EndGlobalSection
 EndGlobal
